refactored to fix build errors
